Added additional about section
