+++ styles/globals.sass
@@ -35,12 +35,17 @@ body
         margin: 0
         padding: 0
 
+    h1
+        font-size: 1.8em
+
+    h2
+        font-size: 1.5em
+
     .pageTitle
         position: relative
         padding-left: 20px
         margin-bottom: 20px
         font-weight: 300
-        font-size: 2em
 
         &:before
             content: ''
